Prevent error from being thrown when creating a console command for a module
On trying to generate a console command for a module, I get the following error after entering the command ID:

`Exception 'yii\base\InvalidArgumentException' with message 'The default value must begin with the base namespace (modules\\site).'`

This is because `$options['default']` is normalized and doesn't contain `\\` anymore, while $this->baseNamespace does. Normalizing the base namespace prevents this from happening.